Tilemap being displayed on screen with references to each tile/layer

+++ unity/Assets/Scripts/Player.cs
@@ -11,9 +11,9 @@ public class Player : FSprite
 
     public Vector2 TileCoordinates { get { return new Vector2(TileX, TileY); } set { tileCoordinates = new Vector2(value.x, value.y); } }
 
-    public float TileX { get { return tileCoordinates.x; } set { tileCoordinates.x = value; } }
+    public int TileX { get { return (int) tileCoordinates.x; } set { tileCoordinates.x = value; } }
 
-    public float TileY { get { return tileCoordinates.y; } set { tileCoordinates.y = value; } }
+    public int TileY { get { return (int) tileCoordinates.y; } set { tileCoordinates.y = value; } }
 
     public Player(string elementName)
         : base(elementName)
